Overhaul1 - It works!
Added main function checkFreq that checks whether it is time to water and then waters if it is.
Added pump function in main that checks whether anyone needs it to be on (pumpSet).
Added if statements that allow serial communication about what is happening. If statements make it so serial.print will only happen once per change not every millisecond that the loop happens.
Did general debugging.